+++ server/Button.js
@@ -1,11 +1,14 @@
 class Button {
-  constructor(id, displayName, phoneNumber, createdAt, updatedAt, buttonSerialNumber, client) {
+  constructor(id, displayName, phoneNumber, createdAt, updatedAt, buttonSerialNumber, isActive, sentLowBatteryAlertAt, sentVitalsAlertAt, client) {
     this.id = id
     this.displayName = displayName
     this.phoneNumber = phoneNumber
     this.createdAt = createdAt
     this.updatedAt = updatedAt
     this.buttonSerialNumber = buttonSerialNumber
+    this.isActive = isActive
+    this.sentLowBatteryAlertAt = sentLowBatteryAlertAt
+    this.sentVitalsAlertAt = sentVitalsAlertAt
     this.client = client
   }
 }